[AIRFLOW-6204] Remove sepearate GCP system tests helper
